# docs/openAPI_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Expense SaaS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/accounts": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create Account",
        "description": "Create a new account owned by the authenticated user.\n\nArgs:\n    payload: AccountCreate schema with account properties.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    The created Account instance.",
        "operationId": "create_account_accounts_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "List Accounts",
        "description": "List accounts belonging to the user and accounts shared with their families (tenants).\n\nArgs:\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    List of serialized account dicts (balance masked according to visibility rules).",
        "operationId": "list_accounts_accounts_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountRead"
                  },
                  "title": "Response List Accounts Accounts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Account",
        "description": "Get a single account by id and serialize according to requestor privileges.\n\nArgs:\n    account_id: UUID of the account.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    Serialized account dict (balance may be masked).\n\nRaises:\n    HTTPException 404 when account does not exist.",
        "operationId": "get_account_accounts__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "accounts"
        ],
        "summary": "Update Account",
        "description": "Update account fields. Only the account owner may modify.\n\nArgs:\n    account_id: UUID of the account to update.\n    payload: AccountUpdate schema with optional fields.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    The updated Account instance.\n\nRaises:\n    HTTPException 404 when account is not found.\n    HTTPException 403 when requester is not the owner.",
        "operationId": "update_account_accounts__account_id__patch",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Delete Account",
        "description": "Delete an account. Only the account owner may delete their account.\n\nArgs:\n    account_id: UUID of the account to delete.\n    db: Async DB session.\n    user: Current authenticated user.\n\nRaises:\n    HTTPException 404 when account does not exist.\n    HTTPException 403 when requester is not the owner.",
        "operationId": "delete_account_accounts__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}/shares": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Account Shares",
        "description": "Return shares for a given account; only the account owner may view shares.\n\nArgs:\n    account_id: UUID of the account.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    List of AccountShare records for the account.\n\nRaises:\n    HTTPException 404 when account not found.\n    HTTPException 403 when requester is not the owner.",
        "operationId": "get_account_shares_accounts__account_id__shares_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountShareRead"
                  },
                  "title": "Response Get Account Shares Accounts  Account Id  Shares Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create Account Share",
        "description": "Create an AccountShare linking an account to a tenant (family).\n\nThe account owner may grant visibility to another tenant.",
        "operationId": "create_account_share_accounts__account_id__shares_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountShareCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountShareRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}/shares/{tenant_id}": {
      "patch": {
        "tags": [
          "accounts"
        ],
        "summary": "Update Account Share",
        "description": "Update properties of an account share identified by tenant_id. Only account owner can modify.",
        "operationId": "update_account_share_accounts__account_id__shares__tenant_id__patch",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountShareUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountShareRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Delete Account Share",
        "description": "Delete an account share identified by tenant_id. Only the account owner may delete.",
        "operationId": "delete_account_share_accounts__account_id__shares__tenant_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "currency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Currency"
              },
              {
                "type": "null"
              }
            ],
            "default": "BRL"
          },
          "balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance",
            "default": "0.00"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "AccountCreate",
        "description": "Input schema to create an account.\n\nArgs:\n    name: Account name.\n    type: Account type.\n    currency: Optional currency (defaults to BRL).\n    balance: Optional starting balance."
      },
      "AccountRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "balance": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "user_name",
          "name",
          "type",
          "currency",
          "balance",
          "created_at",
          "updated_at"
        ],
        "title": "AccountRead",
        "description": "Read schema for account data returned by the API.\n\nAttributes:\n    id: Account identifier.\n    user_id: Owner user id.\n    name: Account name.\n    type: Account type.\n    currency: Currency code.\n    balance: Monetary balance or None when masked.\n    created_at: Creation timestamp.\n    updated_at: Last update timestamp."
      },
...
      "AccountShareRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "visibility": {
            "$ref": "#/components/schemas/ShareVisibility"
          },
          "granted_by": {
            "type": "string",
            "format": "uuid",
            "title": "Granted By"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "tenant_id",
          "visibility",
          "granted_by",
          "granted_at"
        ],
        "title": "AccountShareRead",
        "description": "Read schema for account share data returned by the API.\n\nAttributes:\n    id: Share id.\n    account_id: Shared account id.\n    tenant_id: Tenant receiving the share.\n    visibility: Visibility setting for the share.\n    granted_by: User that granted the share.\n    granted_at: Timestamp when the share was granted."
      },
      "AccountShareUpdate": {
        "properties": {
          "visibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareVisibility"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "AccountShareUpdate",
        "description": "Input schema for updating an account share.\n\nArgs:\n    visibility: New visibility setting (optional)."
      },
...
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountType"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Currency"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance"
          }
        },
        "type": "object",
        "title": "AccountUpdate",
        "description": "Input schema for updating an account.\n\nArgs:\n    name: New account name.\n    type: New account type.\n    currency: New currency code.\n    balance: New balance value."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}